Move Scalar to pgrx::array::Scalar
