+++ pgrx/src/array.rs
@@ -380,3 +380,23 @@ impl Toasty for RawArray {
         unsafe { pg_sys::pfree(self.ptr.as_ptr().cast()) }
     }
 }
+
+/// Marker for "simple scalars" in arrays
+///
+/// A Scalar must have:
+/// - A fixed size
+/// - No padding bits
+/// - All bitpatterns are valid
+/// - Postgres runtime handling which respects these properties
+///
+/// This allows for it to be copied from a Rust slice to a Postgres array,
+/// and obtain a Rust slice from a Postgres array if it contains no nulls.
+pub unsafe trait Scalar: Sized + Copy {}
+
+unsafe impl Scalar for f32 {}
+#[cfg(target_pointer_width = "64")]
+unsafe impl Scalar for f64 {}
+unsafe impl Scalar for i8 {}
+unsafe impl Scalar for i16 {}
+unsafe impl Scalar for i32 {}
+unsafe impl Scalar for i64 {}
